added backwards movement added start timer, potential fix for first goal lost error added timout to the goals, if it doesn't move then resends the goal

+++ maze_ws/src/navigation/exploration/src/ClientScope.h
@@ -10,10 +10,10 @@ struct ClientScope
     string yfeedback;
     string zfeedback;
 
-    uint8_t result;
+    int result;
     string textRes;
     // int status;
-    uint8_t status;
+    int status;
     bool resultReceived;
     bool startedGoal;
